Added class CommonUtils. Refact TelegramRealBotClientTest.
